Disable three nodes test on github ci. Move annotation into main package, so that it is usable from contribs
