+++ contribs/distributed-simulation/src/test/java/org/matsim/dsim/DistributedIntegrationTest.java
@@ -128,7 +128,7 @@ public class DistributedIntegrationTest {
 	 */
 	@Test
 	@Order(3)
-	@DisabledOnGitHubWindowsCI
+	@org.matsim.testcases.DisabledOnGitHubWindowsCI
 	void runDistributed() throws ExecutionException, InterruptedException, TimeoutException, IOException {
 
 		int size = 2;
