+++ contribs/distributed-simulation/src/test/java/org/matsim/dsim/ThreeLinkIntegrationTest.java
@@ -128,7 +128,7 @@ public class ThreeLinkIntegrationTest {
 	@Test
 	@Order(2)
 	@Timeout(value = 2, unit = TimeUnit.MINUTES)
-	@DisabledOnGitHubWindowsCI
+	@org.matsim.testcases.DisabledOnGitHubWindowsCI
 	void oneAgentThreeNodes() throws InterruptedException, ExecutionException, TimeoutException {
 
 		var configPath = utils.getPackageInputDirectory() + "three-links-scenario/three-links-config.xml";
@@ -176,7 +176,7 @@ public class ThreeLinkIntegrationTest {
 	@Test
 	@Order(2)
 	@Timeout(value = 2, unit = TimeUnit.MINUTES)
-	@DisabledOnGitHubWindowsCI
+	@org.matsim.testcases.DisabledOnGitHubWindowsCI
 	void oneAgentThreeNodesTwoIterations() throws ExecutionException, InterruptedException, TimeoutException {
 		var configPath = utils.getPackageInputDirectory() + "three-links-scenario/three-links-config.xml";
 		var outputDirectory = utils.getOutputDirectory(); // this also creats the directory
